+++ src/data/BeeBot.txt
@@ -1,4 +1,1 @@
 Event | 0 | bye | 10 September 2024 1600 | 11 September 2024 1600
-ToDo | 0 | hi
-Event | 0 | buy dog | 01 December 2024 | 02 December 2024
-Deadline | 0 | buy cat | 02 December 2024
