+++ src/main/java/deadline/Deadline.java
@@ -1,3 +1,5 @@
+package deadline;
+import task.Task;
 public class Deadline extends Task {
     private String dateTime;
