+++ src/main/java/event/Event.java
@@ -1,3 +1,5 @@
+package event;
+import task.Task;
 public class Event extends Task {
     private String startTime;
     private String endTime;
