+++ src/main/java/task/Task.java
@@ -1,3 +1,5 @@
+package task;
+
 public class Task {
     private boolean done = false;
     private String name;
